+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## 1.2.2
 
 -   Improve validation error styling in Sampling Settings.